clean up tests that mock dimension
